+++ src/glyphs.js
@@ -22,6 +22,7 @@ export const GLYPHS = {
   boot:       ['........', '........', '........', '........', '####....', '.##.....', '.####...', '.####...'],
   cart:       ['........', '........', '........', '########', '#......#', '#......#', '########', '...##...'],
   belt:       ['........', '########', '#..#..#.', '.#..#..#', '.#..#..#', '#..#..#.', '########', '........'],
+  lift:       ['.....#..', '.###.#..', '.###.#..', '######..', '#....#..', '######..', '.....###', '.#..#...'],
   hoist:      ['........', '#......#', '#......#', '##....##', '##....##', '.#....#.', '.#.##.#.', '.######.'],
   furrow:     ['........', '.#...#..', '.##..##.', '.#...#..', '##..##..', '.#...#..', '.##..##.', '.#...#..'],
   die:        ['...#####', '...#...#', '...#.#.#', '#####..#', '#...####', '#.#.#...', '#...#...', '#####...'],
@@ -90,7 +91,7 @@ export const GLYPH_OF = {
   props: ['planks'], net: ['net'], arch: ['arch'], askwizards: ['point'],
   carry: ['sack'], auto: ['lever'], speed: ['swing'], pick: ['pickhead'],
   critchance: ['spark'], critmult: ['cracked'],
-  haulcarry: ['sack'], haulpace: ['boot', 'up'], carter: ['cart'], belt: ['belt'],
+  haulcarry: ['sack'], haulpace: ['boot', 'up'], carter: ['cart'], driver: ['lift'], belt: ['belt'],
   unlockquarry: ['hoist'], unlockfarm: ['furrow'], unlockapothecary: ['pot'], unlockcasino: ['die'],
   unlockshack: ['hut'], unlockouthouse: ['bucket'], unlocktower: ['tower'], unlockscrub: ['fan'],
   // the house and the closet